Module libyml::api

API module for LibYML

Structs

ScalarEventData

Create a SCALAR event.

Functions

yaml_alias_event_initialize⚠

Create an ALIAS event.

yaml_document_add_mapping⚠

Create a MAPPING node and attach it to the document.

yaml_document_add_scalar⚠

Create a SCALAR node and attach it to the document.

yaml_document_add_sequence⚠

Create a SEQUENCE node and attach it to the document.

yaml_document_append_mapping_pair⚠

Add a pair of a key and a value to a MAPPING node.

yaml_document_append_sequence_item⚠

Add an item to a SEQUENCE node.

yaml_document_delete⚠

Delete a YAML document and all its nodes.

yaml_document_end_event_initialize⚠

Create the DOCUMENT-END event.

yaml_document_get_node⚠

Get a node of a YAML document.

yaml_document_get_root_node⚠

Get the root of a YAML document node.

yaml_document_initialize⚠

Create a YAML document.

yaml_document_start_event_initialize⚠

Create the DOCUMENT-START event.

yaml_emitter_delete⚠

Destroy an emitter.

yaml_emitter_initialize⚠

Initialize an emitter.

yaml_emitter_set_break⚠

Set the preferred line break.

yaml_emitter_set_canonical⚠

Set if the output should be in the “canonical” format as in the YAML specification.

yaml_emitter_set_encoding⚠

Set the output encoding.

yaml_emitter_set_indent⚠

Set the indentation increment.

yaml_emitter_set_output⚠

Set a generic output handler.

yaml_emitter_set_output_string⚠

Set a string output.

yaml_emitter_set_unicode⚠

Set if unescaped non-ASCII characters are allowed.

yaml_emitter_set_width⚠

Set the preferred line width. -1 means unlimited.

yaml_event_delete⚠

Free any memory allocated for an event object.

yaml_free⚠

Free memory allocated by yaml_malloc or yaml_realloc.

yaml_malloc⚠

Allocate memory using the system’s malloc function.

yaml_mapping_end_event_initialize⚠

Create a MAPPING-END event.

yaml_mapping_start_event_initialize⚠

Create a MAPPING-START event.

yaml_parser_delete⚠

Destroy a parser.

yaml_parser_initialize⚠

Initialize a parser.

yaml_parser_set_encoding⚠

Set the source encoding.

yaml_parser_set_input⚠

Set a generic input handler.

yaml_parser_set_input_string⚠

Set a string input.

yaml_queue_extend⚠

Extend a queue by reallocating and copying the existing data.

yaml_realloc⚠

Reallocate memory using the system’s realloc function.

yaml_scalar_event_initialize⚠

Create a SCALAR event.

yaml_sequence_end_event_initialize⚠

Create a SEQUENCE-END event.

yaml_sequence_start_event_initialize⚠

Create a SEQUENCE-START event.

yaml_stack_extend⚠

Extend a stack by reallocating and copying the existing data.

yaml_strdup⚠

Duplicate a string using the system’s strdup function.

yaml_stream_end_event_initialize⚠

Create the STREAM-END event.

yaml_stream_start_event_initialize⚠

Create the STREAM-START event.

yaml_string_extend⚠

Extend a string buffer by reallocating and copying the existing data.

yaml_string_join⚠

Join two string buffers by copying data from one to the other.

yaml_token_delete⚠

Free any memory allocated for a token object.