libyml::api
Function yaml_emitter_set_break

pub unsafe fn yaml_emitter_set_break(
    emitter: *mut YamlEmitterT,
    line_break: YamlBreakT,
)

Set the preferred line break.

This function sets the preferred line break character to be used by the emitter.

Safety

emitter must be a valid, non-null pointer to a properly initialized YamlEmitterT struct.

The YamlEmitterT struct and its associated data structures must be properly aligned and have the expected memory layout.