libyml::api
Function yaml_parser_set_input

pub unsafe fn yaml_parser_set_input(
    parser: *mut YamlParserT,
    handler: YamlReadHandlerT,
    data: *mut c_void,
)

Set a generic input handler.

This function sets a custom input handler for the parser.

Safety

parser must be a valid, non-null pointer to a properly initialized YamlParserT struct.

The YamlParserT struct must not have an input handler already set.

handler must be a valid function pointer that follows the signature of YamlReadHandlerT.

data must be a valid pointer that will be passed to the handler function.

The YamlParserT struct and its associated data structures must be properly aligned and have the expected memory layout.