Function libyml::api::yaml_stream_start_event_initialize

pub unsafe fn yaml_stream_start_event_initialize(
    event: *mut YamlEventT,
    encoding: YamlEncodingT
) -> Success

Create the STREAM-START event.

This function initializes a YamlEventT struct with the type YamlStreamStartEvent. It is used to signal the start of a YAML stream being emitted.

Safety

event must be a valid, non-null pointer to a YamlEventT struct that can be safely written to.

The YamlEventT struct must be properly aligned and have the expected memory layout.