libyml
Module libc

Declarations for C library functions used within the LibYML library.

This module contains the necessary types and constants required for interfacing with C libraries, particularly in the context of memory management and low-level operations within LibYML.

Re-exports

pub use core::primitive::i32 as c_int;

pub use core::primitive::i64 as c_long;

pub use core::primitive::i8 as c_char;

pub use core::primitive::u32 as c_uint;

pub use core::primitive::u64 as c_ulong;

pub use core::primitive::u8 as c_uchar;

Enums

c_void

Equivalent to C’s void type when used as a pointer.