libyml
Macro IS_HEX_AT

macro_rules! IS_HEX_AT {
    ($string:expr, $offset:expr) => { ... };
}

Checks if the character at the given offset in the string is a hexadecimal digit.

Parameters

string: A reference to the string to check.

offset: The offset in the string to check.

Return

Returns true if the character at the given offset is a hexadecimal digit (0-9, A-F, a-f), and false otherwise.