Function libyml::success::is_failure

pub fn is_failure(result: Success) -> bool

Checks if the given Success result indicates a failure operation.

Arguments

result - A Success struct representing the result of an operation.

Returns

true if the operation failed.

false if the operation was successful.