libyml::success
Function is_success

pub fn is_success(result: Success) -> bool

Checks if the given Success result indicates a successful operation.

Arguments

result - A Success struct representing the result of an operation.

Returns

true if the operation was successful.

false if the operation failed.