Module libyml::yaml

YAML API module for LibYML.

This module provides functions and types for working directly with YAML data structures.

Re-exports

pub use core::primitive::i64 as ptrdiff_t;

pub use core::primitive::u64 as size_t;

pub use core::primitive::u8 as yaml_char_t;

Structs

UnnamedYamlDocumentTTagDirectives

Represents the list of tag directives in a YAML document.

UnnamedYamlEmitterTOutput

Represents the output data associated with a YAML emitter.

UnnamedYamlEmitterTOutputString

Represents the unamed output string data associated with a YAML emitter.

UnnamedYamlEventTData

Represents the data associated with a YAML event.

UnnamedYamlEventTdataAlias

Represents the data associated with a YAML alias event.

UnnamedYamlEventTdataDocumentEnd

Represents the data associated with the end of a YAML document.

UnnamedYamlEventTdataDocumentStart

Represents the data associated with the start of a YAML document.

UnnamedYamlEventTdataDocumentStartTagDirectives

Represents the list of tag directives at the start of a YAML document.

UnnamedYamlEventTdataMappingStart

Represents the data associated with the start of a YAML mapping.

UnnamedYamlEventTdataScalar

Represents the data associated with a YAML scalar event.

UnnamedYamlEventTdataSequenceStart

Represents the data associated with the start of a YAML sequence.

UnnamedYamlEventTdataStreamStart

Represents the data associated with the start of a YAML stream.

UnnamedYamlNodeTData

Represents the data associated with a YAML node.

UnnamedYamlNodeTDataMapping

Represents the data associated with a YAML mapping node.

UnnamedYamlNodeTDataScalar

Represents the data associated with a YAML scalar node.

UnnamedYamlNodeTDataSequence

Represents the data associated with a YAML sequence node.

UnnamedYamlTokenTData

The data structure for YAML tokens.

UnnamedYamlTokenTdataAlias

Represents an alias in a YAML document.

UnnamedYamlTokenTdataAnchor

Represents an anchor in a YAML document.

UnnamedYamlTokenTdataScalar

Represents a scalar value in a YAML document.

UnnamedYamlTokenTdataStreamStart

Represents the start of a YAML data stream.

UnnamedYamlTokenTdataTag

Represents a tag in a YAML document.

UnnamedYamlTokenTdataTagDirective

Represents the tag directive in a YAML document.

UnnamedYamlTokenTdataVersionDirective

Represents the version directive in a YAML document.

YamlAliasDataT

This structure holds aliases data.

YamlBufferT

Represents the data associated with a YAML token.

YamlDocumentT

The document structure.

YamlEmitterT

The emitter structure.

YamlEmitterTPrefix

Represents the prefix data associated with a YAML emitter.

YamlEventT

The event structure.

YamlMarkT

The pointer position.

YamlNodePairT

An element of a mapping node.

YamlNodeT

The node structure.

YamlParserT

The parser structure.

YamlParserTPrefix

Represents the prefix data associated with a YAML parser.

YamlSimpleKeyT

This structure holds information about a potential simple key.

YamlStackT

The beginning of the stack.

YamlTagDirectiveT

The tag directive data.

YamlTokenT

The token structure.

YamlVersionDirectiveT

The version directive data.

Enums

YamlBreakT

Line break type.

YamlEmitterStateT

The emitter states.

YamlEncodingT

The stream encoding.

YamlErrorTypeT

Many bad things could happen with the parser and emitter.

YamlEventTypeT

Event types.

YamlMappingStyleT

Mapping styles.

YamlNodeTypeT

Node types.

YamlParserStateT

The states of the parser.

YamlScalarStyleT

Scalar styles.

YamlSequenceStyleT

Sequence styles.

YamlTokenTypeT

The token types.

Type Aliases

YamlNodeItemT

Represents an element of a YAML sequence node.

YamlReadHandlerT

The prototype of a read handler.

YamlWriteHandlerT

The prototype of a write handler.