Type Alias libyml::yaml::YamlReadHandlerT

pub type YamlReadHandlerT = unsafe fn(data: *mut c_void, buffer: *mut u8, size: u64, size_read: *mut u64) -> i32;

The prototype of a read handler.

The read handler is called when the parser needs to read more bytes from the source. The handler should write not more than size bytes to the buffer. The number of written bytes should be set to the length variable.

On success, the handler should return 1. If the handler failed, the returned value should be 0. On EOF, the handler should set the size_read to 0 and return 1.