libyml::yaml
Type Alias YamlWriteHandlerT

pub type YamlWriteHandlerT = unsafe fn(data: *mut c_void, buffer: *mut u8, size: u64) -> i32;

The prototype of a write handler.

The write handler is called when the emitter needs to flush the accumulated characters to the output. The handler should write size bytes of the buffer to the output.

On success, the handler should return 1. If the handler failed, the returned value should be 0.