libyml

Module success

Success and Failure types for LibYML.

This module provides types for representing the success and failure of various operations within the library.

Structs

Failure

Structure representing the failure state of an operation.

Success

Structure representing the success state of an operation.

Constants

FAIL

Constant representing a failed operation.

OK

Constant representing a successful operation.

Functions

is_failure

Checks if the given Success result indicates a failure operation.

is_success

Checks if the given Success result indicates a successful operation.