Module libyml::yaml

Structs

UnnamedYamlDocumentTTagDirectives

UnnamedYamlEventTData

UnnamedYamlEventTdataAlias

UnnamedYamlEventTdataDocumentEnd

UnnamedYamlEventTdataDocumentStart

UnnamedYamlEventTdataDocumentStartTagDirectives

UnnamedYamlEventTdataMappingStart

UnnamedYamlEventTdataScalar

UnnamedYamlEventTdataSequenceStart

UnnamedYamlEventTdataStreamStart

UnnamedYamlNodeTData

UnnamedYamlNodeTDataMapping

UnnamedYamlNodeTDataScalar

UnnamedYamlNodeTDataSequence

UnnamedYamlTokenTData

UnnamedYamlTokenTdataAlias

UnnamedYamlTokenTdataAnchor

UnnamedYamlTokenTdataScalar

UnnamedYamlTokenTdataStreamStart

UnnamedYamlTokenTdataTag

UnnamedYamlTokenTdataTagDirective

UnnamedYamlTokenTdataVersionDirective

YamlAliasDataT

This structure holds aliases data.

YamlDocumentT

The document structure.

YamlEmitterT

The emitter structure.

YamlEmitterTPrefix

YamlEventT

The event structure.

YamlMarkT

The pointer position.

YamlNodePairT

An element of a mapping node.

YamlNodeT

The node structure.

YamlParserT

The parser structure.

YamlParserTPrefix

YamlSimpleKeyT

This structure holds information about a potential simple key.

YamlStackT

The beginning of the stack.

YamlTagDirectiveT

The tag directive data.

YamlTokenT

The token structure.

YamlVersionDirectiveT

The version directive data.

Enums

YamlBreakT

Line break type.

YamlEmitterStateT

The emitter states.

YamlEncodingT

The stream encoding.

YamlErrorTypeT

Many bad things could happen with the parser and emitter.

YamlEventTypeT

Event types.

YamlMappingStyleT

Mapping styles.

YamlNodeTypeT

Node types.

YamlParserStateT

The states of the parser.

YamlScalarStyleT

Scalar styles.

YamlSequenceStyleT

Sequence styles.

YamlTokenTypeT

Token types.

Type Aliases

YamlNodeItemT

An element of a sequence node.

YamlReadHandlerT

The prototype of a read handler.

YamlWriteHandlerT

The prototype of a write handler.